requirements.txt
gunicorn
numpy
pandas
scikit-learn
streamlit
geopanda
plotly